+++ src/module.mli
@@ -170,4 +170,4 @@ val set_src_dir : t -> src_dir:Path.t -> t
 val generated : src_dir:Path.t -> Name.t -> t
 
 (** Represent the generated alias module. *)
-val generated_alias : src_dir:Path.t -> Name.t -> t
+val generated_alias : src_dir:Path.Build.t -> Name.t -> t